MVC Javascript bundler is apparently retarded
Renamed jqColorPicker so we can make changes to it if necessary. The
bundler was pulling bunnies out of magic hats if I used the name
"jqColorPicker". I literally completely emptied the file, and it still
somehow published code for that file... Complete pain in the ass piece
of junk stupid *#$*#*$(*#$(*#*$$(